+++ src/lib/phone-input-core/index.ts
@@ -1,5 +1,5 @@
 export { DEFAULT_LABELS, type PhoneInputLabels, type PickerCloseReason } from "./labels";
-export { RegionFlag } from "./region-flag";
+export { flagEmoji, RegionFlag } from "./region-flag";
 export {
   type PhoneMaskState,
   type PhoneValue,